TargetClass
Targetclass plus vähän siistimistä
