--- Cyber Wizards 3D/Assets/DrawCircle.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: a982cbde783ea3c4ea97a2b6ae9bc2df
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
