--- Cyber Wizards 3D/Assets/RandomPointOnNavMesh.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 36b34c3e97f1a6d47b6fd2a81369f2c1
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
