+++ Cyber Wizards 3D/Assets/Scripts/PlayerInput.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: fd5b75826f66b5c45815678d81d2f7d6
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
