+++ Cyber Wizards 3D/Assets/Scripts/StatsUI.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 8a98e90d34b4bad48a430beec5c4229e
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
