+++ Cyber Wizards 3D/Assets/Scripts/TargetClass.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 27a5d8086d04b9a469b9784dd7257565
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
