+++ Cyber Wizards 3D/Assets/Scripts/Turn.cs
@@ -1,0 +1,8 @@
+﻿using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+public enum Turn {
+	ON,
+	OFF
+};
