+++ Cyber Wizards 3D/Assets/Scripts/TurnManager.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 9f1b6a86652eb654abe1e5d62ce1bbdc
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
