+++ Cyber Wizards 3D/Assets/Scripts/TurnStatus.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: da0147f6d808dd94b928f08f6f881f67
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
