--- Cyber Wizards 3D/Assets/Spell.cs.meta
@@ -1,13 +1,0 @@
-fileFormatVersion: 2
-guid: 6159bc1e8580ede44838b17d9445a0c7
-timeCreated: 1520592401
-licenseType: Pro
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
