--- Cyber Wizards 3D/Assets/Turn.cs
@@ -1,8 +1,0 @@
-﻿using System.Collections;
-using System.Collections.Generic;
-using UnityEngine;
-
-public enum Turn {
-	ON,
-	OFF
-};
