--- Cyber Wizards 3D/Assets/Turn.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: f25601f35c33b0c4db2ba78bd65b99c9
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
